Changed data type from long to BigDecimal for avgRequestPerSec (#4)

+++ domain-core/src/main/kotlin/tech/testra/reportal/domain/valueobjects/ValueObjects.kt
@@ -44,7 +44,7 @@ data class SimulationScenario(
     val percentiles: List<Percentile>,
     val average: BigDecimal,
     val stdDiv: BigDecimal,
-    val avgRequestPerSec: Long
+    val avgRequestPerSec: BigDecimal
 )
 
 data class Percentile(
